Refactor functional test base structs
