+++ tests/testcore/client_functional_suite.go
@@ -45,16 +45,16 @@ import (
 
 type (
 	ClientFunctionalSuite struct {
+		FunctionalTestBase
+
 		// override suite.Suite.Assertions with require.Assertions; this means that s.NotNil(nil) will stop the test,
 		// not merely log an error
 		*require.Assertions
-		FunctionalTestBase
 		historyrequire.HistoryRequire
+
 		sdkClient sdkclient.Client
 		worker    worker.Worker
 		taskQueue string
-
-		baseConfigPath string
 	}
 )
